@@ -7,7 +7,7 @@ repository: https://github.com/daniel-maxhari/dynamic_text_highlighting.git
 issue_tracker: https://github.com/daniel-maxhari/dynamic_text_highlighting/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
